API/drzz_services/requirements.txt
Flask
gunicorn
google-cloud-datastore
bcrypt
google-cloud-storage

